requirements.txt
websockets
paho-mqtt
pyserial
